requirements.txt
flask
psycopg2-binary
python-dotenv
openai
google-generativeai